[REF] Add gscontrol module (#240)
* Move gscontrol_raw and gscontrol_mmix into new gscontrol module.

* Make bracket pretty.
