+++ tedana/model/fit.py
@@ -7,7 +7,6 @@ import os.path as op
 import numpy as np
 import pandas as pd
 from scipy import stats
-from scipy.special import lpmv
 import nilearn.image as niimg
 from nilearn._utils import check_niimg
 from nilearn.regions import connected_regions
@@ -406,94 +405,6 @@ def get_coeffs(data, X, mask=None, add_const=False):
         betas = utils.unmask(betas, mask)
 
     return betas
-
-
-def gscontrol_raw(catd, optcom, n_echos, ref_img, dtrank=4):
-    """
-    Removes global signal from individual echo `catd` and `optcom` time series
-
-    This function uses the spatial global signal estimation approach to
-    to removal global signal out of individual echo time series datasets. The
-    spatial global signal is estimated from the optimally combined data after
-    detrending with a Legendre polynomial basis of `order = 0` and
-    `degree = dtrank`.
-
-    Parameters
-    ----------
-    catd : (S x E x T) array_like
-        Input functional data
-    optcom : (S x T) array_like
-        Optimally combined functional data (i.e., the output of `make_optcom`)
-    n_echos : :obj:`int`
-        Number of echos in data. Should be the same as `E` dimension of `catd`
-    ref_img : :obj:`str` or img_like
-        Reference image to dictate how outputs are saved to disk
-    dtrank : :obj:`int`, optional
-        Specifies degree of Legendre polynomial basis function for estimating
-        spatial global signal. Default: 4
-
-    Returns
-    -------
-    dm_catd : (S x E x T) array_like
-        Input `catd` with global signal removed from time series
-    dm_optcom : (S x T) array_like
-        Input `optcom` with global signal removed from time series
-    """
-    LGR.info('Applying amplitude-based T1 equilibration correction')
-    if catd.shape[0] != optcom.shape[0]:
-        raise ValueError('First dimensions of catd ({0}) and optcom ({1}) do not '
-                         'match'.format(catd.shape[0], optcom.shape[0]))
-    elif catd.shape[1] != n_echos:
-        raise ValueError('Second dimension of catd ({0}) does not match '
-                         'n_echos ({1})'.format(catd.shape[1], n_echos))
-    elif catd.shape[2] != optcom.shape[1]:
-        raise ValueError('Third dimension of catd ({0}) does not match '
-                         'second dimension of optcom '
-                         '({1})'.format(catd.shape[2], optcom.shape[1]))
-
-    # Legendre polynomial basis for denoising
-    bounds = np.linspace(-1, 1, optcom.shape[-1])
-    Lmix = np.column_stack([lpmv(0, vv, bounds) for vv in range(dtrank)])
-
-    # compute mean, std, mask local to this function
-    # inefficient, but makes this function a bit more modular
-    Gmu = optcom.mean(axis=-1)  # temporal mean
-    Gmask = Gmu != 0
-
-    # find spatial global signal
-    dat = optcom[Gmask] - Gmu[Gmask][:, np.newaxis]
-    sol = np.linalg.lstsq(Lmix, dat.T, rcond=None)[0]  # Legendre basis for detrending
-    detr = dat - np.dot(sol.T, Lmix.T)[0]
-    sphis = (detr).min(axis=1)
-    sphis -= sphis.mean()
-    io.filewrite(utils.unmask(sphis, Gmask), 'T1gs', ref_img)
-
-    # find time course ofc the spatial global signal
-    # make basis with the Legendre basis
-    glsig = np.linalg.lstsq(np.atleast_2d(sphis).T, dat, rcond=None)[0]
-    glsig = stats.zscore(glsig, axis=None)
-    np.savetxt('glsig.1D', glsig)
-    glbase = np.hstack([Lmix, glsig.T])
-
-    # Project global signal out of optimally combined data
-    sol = np.linalg.lstsq(np.atleast_2d(glbase), dat.T, rcond=None)[0]
-    tsoc_nogs = dat - np.dot(np.atleast_2d(sol[dtrank]).T,
-                             np.atleast_2d(glbase.T[dtrank])) + Gmu[Gmask][:, np.newaxis]
-
-    io.filewrite(optcom, 'tsoc_orig', ref_img)
-    dm_optcom = utils.unmask(tsoc_nogs, Gmask)
-    io.filewrite(dm_optcom, 'tsoc_nogs', ref_img)
-
-    # Project glbase out of each echo
-    dm_catd = catd.copy()  # don't overwrite catd
-    for echo in range(n_echos):
-        dat = dm_catd[:, echo, :][Gmask]
-        sol = np.linalg.lstsq(np.atleast_2d(glbase), dat.T, rcond=None)[0]
-        e_nogs = dat - np.dot(np.atleast_2d(sol[dtrank]).T,
-                              np.atleast_2d(glbase.T[dtrank]))
-        dm_catd[:, echo, :] = utils.unmask(e_nogs, Gmask)
-
-    return dm_catd, dm_optcom
 
 
 def spatclust(img, min_cluster_size, threshold=None, index=None, mask=None):
