+++ tedana/tests/test_gscontrol.py
@@ -5,7 +5,7 @@ Tests for tedana.model.fit
 import numpy as np
 import pytest
 
-from tedana.model import fit
+import tedana.gscontrol as gsc
 
 
 def test_break_gscontrol_raw():
@@ -20,14 +20,14 @@ def test_break_gscontrol_raw():
 
     catd = np.empty((n_samples+1, n_echos, n_vols))
     with pytest.raises(ValueError) as e_info:
-        fit.gscontrol_raw(catd=catd, optcom=optcom, n_echos=n_echos,
+        gsc.gscontrol_raw(catd=catd, optcom=optcom, n_echos=n_echos,
                           ref_img=ref_img, dtrank=4)
     assert str(e_info.value) == ('First dimensions of catd ({0}) and optcom ({1}) do not '
                                  'match'.format(catd.shape[0], optcom.shape[0]))
 
     catd = np.empty((n_samples, n_echos+1, n_vols))
     with pytest.raises(ValueError) as e_info:
-        fit.gscontrol_raw(catd=catd, optcom=optcom, n_echos=n_echos,
+        gsc.gscontrol_raw(catd=catd, optcom=optcom, n_echos=n_echos,
                           ref_img=ref_img, dtrank=4)
     assert str(e_info.value) == ('Second dimension of catd ({0}) does not match '
                                  'n_echos ({1})'.format(catd.shape[1], n_echos))
@@ -35,7 +35,7 @@ def test_break_gscontrol_raw():
     catd = np.empty((n_samples, n_echos, n_vols))
     optcom = np.empty((n_samples, n_vols+1))
     with pytest.raises(ValueError) as e_info:
-        fit.gscontrol_raw(catd=catd, optcom=optcom, n_echos=n_echos,
+        gsc.gscontrol_raw(catd=catd, optcom=optcom, n_echos=n_echos,
                           ref_img=ref_img, dtrank=4)
     assert str(e_info.value) == ('Third dimension of catd ({0}) does not match '
                                  'second dimension of optcom '
